Refactor DisputeClient to use pydantic types
