+++ tests/test_paystack.py
@@ -34,7 +34,7 @@ from pypaystack2.sub_clients.async_clients.customers import AsyncCustomerClient
 from pypaystack2.sub_clients.async_clients.dedicated_accounts import (
     AsyncDedicatedAccountClient,
 )
-from pypaystack2.sub_clients.disputes import AsyncDisputeClient
+from pypaystack2.sub_clients.async_clients.disputes import AsyncDisputeClient
 from pypaystack2.sub_clients.integration import AsyncIntegrationClient
 from pypaystack2.sub_clients.miscellaneous import AsyncMiscellaneousClient
 from pypaystack2.sub_clients.payment_pages import AsyncPaymentPageClient
